requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
distlib==0.3.0
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.1
isort==4.3.21
kafka-python==2.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mirakuru==2.3.0
more-itertools==8.4.0
packaging==20.4
pipenv==2020.6.2
pluggy==0.13.1
port-for==0.4
protobuf==3.12.2
psutil==5.7.0
psycopg2==2.8.5
py==1.8.2
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
pytest-postgresql==2.4.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-mock==1.8.0
six==1.14.0
termcolor==1.1.0
toml==0.10.1
tox==3.15.2
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.23
virtualenv-clone==0.5.4
wcwidth==0.2.4
wrapt==1.12.1
zipp==3.1.0
